Add basic fetch error handling and blank-slate for no issue results
The search API doesn't say whether or not you have access to the repos you're
trying to search, it just returns empty results if the repo doesn't exist or if
you don't have access. Not a simple way to programmatically test this without
explicit additional API calls, so just going with a "huh, no results, here's
what the problem might be" message.